[REF] Make Excel tests more usable by other modules by putting more functionality into Common
