+++ pattern_import_export_xlsx/tests/test_pattern_import.py
@@ -3,7 +3,6 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 import base64
 from io import BytesIO
-from os import path
 
 # TODO FIXME somehow Travis complains that openpyxl isn't there,
 # the warning shows only here and not in any other import of openpyxl?
@@ -13,67 +12,16 @@ import openpyxl
 from odoo.tests import SavepointCase
 from odoo.tools import mute_logger
 
-# helper to dump the result of the import into an excel file
-DUMP_OUTPUT = False
+from .common import ExportPatternExcelCommon
 
 
-PATH = path.dirname(__file__) + "/fixtures/"
-
-
-class TestPatternImport(SavepointCase):
-    @classmethod
-    def setUpClass(cls):
-        super().setUpClass()
-        cls.env = cls.env(
-            context=dict(
-                cls.env.context, test_queue_job_no_delay=True  # no jobs thanks
-            )
-        )
-        cls.ir_export_partner = cls.env["ir.exports"].create(
-            {
-                "name": "Partner",
-                "resource": "res.partner",
-                "is_pattern": True,
-                "export_format": "xlsx",
-            }
-        )
-        cls.ir_export_users = cls.env["ir.exports"].create(
-            {
-                "name": "User",
-                "resource": "res.users",
-                "is_pattern": True,
-                "export_format": "xlsx",
-            }
-        )
-        cls.user_admin = cls.env.ref("base.user_admin")
-        cls.user_demo = cls.env.ref("base.user_demo")
-
-    @classmethod
-    def _load_file(cls, filename, export_id):
-        data = base64.b64encode(open(PATH + filename, "rb").read())
-        wizard = cls.env["import.pattern.wizard"].create(
-            {
-                "ir_exports_id": export_id.id,
-                "import_file": data,
-                "filename": "example.xlsx",
-            }
-        )
-        wizard.action_launch_import()
-
-        if DUMP_OUTPUT:
-            attachment = cls.env["patterned.import.export"].search(
-                [], limit=1, order="id desc"
-            )
-            output_name = filename.replace(".xlsx", ".result.xlsx")
-            with open(output_name, "wb") as output:
-                output.write(base64.b64decode(attachment.datas))
-
+class TestPatternImport(ExportPatternExcelCommon, SavepointCase):
     def test_import_partners_ok(self):
         """
         * Lookup by email
         * Update some o2m fields
         """
-        self._load_file("example.partners.ok.xlsx", self.ir_export_partner)
+        self._load_excel_file("example.partners.ok.xlsx", self.ir_export_partner)
         # check first line
         partner = self.env.ref("base.res_partner_1")
 
@@ -137,7 +85,7 @@ class TestPatternImport(SavepointCase):
         * Lookup by email
         * Report error in excel file through wrong email
         """
-        self._load_file("example.partners.fail.xlsx", self.ir_export_partner)
+        self._load_excel_file("example.partners.fail.xlsx", self.ir_export_partner)
         self.env.clear()
 
         # check that nothong have been done
@@ -174,7 +122,7 @@ class TestPatternImport(SavepointCase):
         * Lookup by DB ID
         * Simple update
         """
-        self._load_file("example.users.ok.xlsx", self.ir_export_users)
+        self._load_excel_file("example.users.ok.xlsx", self.ir_export_users)
         self.assertEqual(self.user_admin.name, "Mitchell Admin Updated")
         self.assertEqual(self.user_demo.name, "Marc Demo Updated")
 
@@ -185,7 +133,7 @@ class TestPatternImport(SavepointCase):
         * Simple update
         """
         self.ir_export_users.use_description = True
-        self._load_file("example.users.descriptive.ok.xlsx", self.ir_export_users)
+        self._load_excel_file("example.users.descriptive.ok.xlsx", self.ir_export_users)
         self.assertEqual(self.user_admin.name, "Mitchell Admin Updated")
         self.assertEqual(self.user_demo.name, "Marc Demo Updated")
 
@@ -196,7 +144,7 @@ class TestPatternImport(SavepointCase):
         * Lookup by external ID
         * Report error in excel file through external id not found
         """
-        self._load_file("example.users.fail.xlsx", self.ir_export_users)
+        self._load_excel_file("example.users.fail.xlsx", self.ir_export_users)
         attachment = self.env["ir.attachment"].search([], order="id desc", limit=1)
         infile = BytesIO(base64.b64decode(attachment.datas))
         wb = openpyxl.load_workbook(filename=infile)
@@ -206,7 +154,7 @@ class TestPatternImport(SavepointCase):
         self.assertTrue(ws["A3"].value)
 
     def test_import_partners_with_parents(self):
-        self._load_file("example.partners.parent.xlsx", self.ir_export_partner)
+        self._load_excel_file("example.partners.parent.xlsx", self.ir_export_partner)
         partner_parent = self.env["res.partner"].search([("name", "=", "Apple")])
         self.assertTrue(partner_parent)
         partner_child = self.env["res.partner"].search([("name", "=", "Steve Jobs")])
